fix(many): refine zoom and add animation, and more refactoring
