+++ package.json
@@ -19,9 +19,11 @@
     "cssnano": "^3.10.0",
     "d3": "^4.9.1",
     "d3-array": "^1.2.0",
+    "d3-ease": "^1.0.5",
     "d3-force": "^1.1.0",
     "d3-scale": "^1.0.0",
     "d3-selection": "^1.1.0",
+    "d3-transition": "^1.1.3",
     "d3-zoom": "^1.7.3",
     "fast-csv": "^2.4.0",
     "file-saver": "^1.3.3",
