+++ src/DataDictionary/graph/GraphCalculator/graphCalculatorHelper.test.js
@@ -5,15 +5,6 @@ import {
 import {
   calculateGraphLayout,
   getAllTypes,
-  getAllChildrenNodeIDs,
-  getAllChildrenLinks,
-  isArticulationNodeInSubgraph,
-  getArticulationNodesInSubgraph,
-  BFSTraverseSubgraph,
-  sortNodesByTopology,
-  getNodesAndLinksSummaryBetweenNodesInSubgraph,
-  getNodeIDsThatHaveNoInLinks,
-  getNodeIDsThatHaveNoOutLinks,
   calculateHighlightRelatedNodeIDs,
   calculatePathRelatedToSecondHighlightingNode,
   calculateDataModelStructure,
@@ -36,81 +27,6 @@ describe('graphCalculatorHelper', () => {
   it('can get all types', () => {
     const types = getAllTypes(testGraph1.graphNodes);
     expect(types).toEqual(testGraph1.expectedNodeTypes);
-  });
-
-  it('can get all children node IDs', () => {
-    const resultNodeIDs = getAllChildrenNodeIDs(testGraph1.startingNode, testGraph1.graphNodes);
-    expect(resultNodeIDs).toEqual(testGraph1.expectedChildrenNodeIDs);
-  });
-
-  it('can get all children links', () => {
-    const resultLinks = getAllChildrenLinks(testGraph1.startingNode, testGraph1.graphNodes);
-    expect(resultLinks).toEqual(testGraph1.expectedChildrenLinks);
-  });
-
-  it('can judge articulation node from subgraph', () => {
-    const result = isArticulationNodeInSubgraph(
-      testGraph1.expectedArticulationNodesInSubgraph[0],
-      testGraph1.testSubgraph,
-      testGraph1.testSubgraphEdges,
-      testGraph1.graphNodes,
-    );
-    expect(result).toBe(true);
-  });
-
-  it('can get articulation nodes from subgraph', () => {
-    const result = getArticulationNodesInSubgraph(
-      testGraph1.startingNode.id,
-      testGraph1.testSubgraph,
-      testGraph1.testSubgraphEdges,
-      testGraph1.graphNodes,
-    );
-    expect(result).toEqual(testGraph1.expectedArticulationNodesInSubgraph);
-  });
-
-  it('can averse subgraph in BFS order', () => {
-    const bfsResult = BFSTraverseSubgraph(
-      testGraph1.testSubgraph,
-      testGraph1.testSubgraphEdges,
-      testGraph1.graphNodes,
-    );
-    expect(bfsResult).toEqual(testGraph1.expectedBFSTraverseSubgraph);
-  });
-
-  it('can sort nodes in subgraph in topology order', () => {
-    const sorted = sortNodesByTopology(
-      testGraph1.testNodeIDsForSort,
-      testGraph1.testSubgraph,
-      testGraph1.testSubgraphEdges,
-      testGraph1.graphNodes,
-    );
-    expect(sorted).toEqual(testGraph1.expectedSorteddNodeIDs);
-  });
-
-  it('can get nodes and links summary between nodes in subgraph', () => {
-    const result = getNodesAndLinksSummaryBetweenNodesInSubgraph(
-      testGraph1.testNode1,
-      testGraph1.testNode2,
-      testGraph1.testSubgraph,
-      testGraph1.testSubgraphEdges,
-      testGraph1.graphNodes,
-    );
-    expect(result).toEqual(testGraph1.expectedSummary);
-  });
-
-  it('can get nodes that have no in/out links in subgraph', () => {
-    const resultNodeIDsWithoutInLinks = getNodeIDsThatHaveNoInLinks(
-      testGraph1.testSubgraph,
-      testGraph1.testSubgraphEdges,
-      testGraph1.graphNodes,
-    );
-    const resultNodeIDsWithoutOutLinks = getNodeIDsThatHaveNoOutLinks(
-      testGraph1.testSubgraph,
-      testGraph1.testSubgraphEdges,
-      testGraph1.graphNodes,
-    );
-    expect(resultNodeIDsWithoutInLinks).toEqual(testGraph1.expectedNodeIDsWithNoInLinks);
-    expect(resultNodeIDsWithoutOutLinks).toEqual(testGraph1.expectedNodeIDsWithNoOutLinks);
   });
 
   it('can calculate related highlighting node IDs', () => {
