+++ src/Login/ProtectedContent.css
@@ -1,0 +1,12 @@
+.protected-content {
+  width: 1220px;
+  margin: auto;
+  min-height: calc(100vh - 221px);
+  background-color: var(--g3-color__bg-cloud);
+}
+
+.protected-content--full-screen {
+  width: 100%;
+  min-width: 1220px;
+  height: calc(100vh - 121px);
+}
